Support the case that the Docker socket is not accessible by the current user
On some of my systems, I don't want to allow the non-root user to access
the Docker socket for security reasons. This is now supported by
creating a file in sudoers.d which allows the user to start the
transmission-nordvpn containers.

Also supports using 'docker-compose' (which is available in the official
Debian sources) instead of 'docker compose' (the compose subcommand is
currently not available in the docker version in the official Debian
sources).